move getMedia selector to ai image hook, use selector/useEffect combo to try and get the value
